+++ localflavor/mx/forms.py
@@ -118,9 +118,9 @@ class MXRFCField(RegexField):
         value = value.upper()
         if self._has_homoclave(value):
             if not value[-1] == self._checksum(value[:-1]):
-                raise ValidationError(self.error_messages['invalid_checksum'])
+                raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
         if self._has_inconvenient_word(value):
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
         return value
 
     def _has_homoclave(self, rfc):
@@ -198,9 +198,9 @@ class MXCLABEField(RegexField):
         if value in self.empty_values:
             return self.empty_value
         if not value.isdigit():
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
         if not self._checksum(value):
-            raise ValidationError(self.error_messages['invalid_checksum'])
+            raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
 
         return value
 
@@ -250,9 +250,9 @@ class MXCURPField(RegexField):
             return self.empty_value
         value = value.upper()
         if value[-1] != self._checksum(value[:-1]):
-            raise ValidationError(self.error_messages['invalid_checksum'])
+            raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
         if self._has_inconvenient_word(value):
-            raise ValidationError(self.error_messages['invalid'])
+            raise ValidationError(self.error_messages['invalid'], code='invalid')
         return value
 
     def _checksum(self, value):
@@ -307,7 +307,7 @@ class MXSocialSecurityNumberField(RegexField):
         if value in self.empty_values:
             return self.empty_value
         if value[-1] != self.__checksum(value[:-1]):
-            raise ValidationError(self.error_messages['invalid_checksum'])
+            raise ValidationError(self.error_messages['invalid_checksum'], code='invalid_checksum')
         return value
 
     def __checksum(self, value):